Split schema backup CLI from API and move under karapace.backup
