+++ setup.py
@@ -46,7 +46,7 @@ setup(
     entry_points={
         "console_scripts": [
             "karapace = karapace.karapace_all:main",
-            "karapace_schema_backup = karapace.schema_backup:main",
+            "karapace_schema_backup = karapace.backup.cli:main",
             "karapace_mkpasswd = karapace.auth:main",
         ],
     },
